Fixed apis folder and refactored import apis
