+++ app/controllers/api/public/articles_controller.rb
@@ -1,8 +1,13 @@
 # frozen_string_literal: true
 
 class Api::Public::ArticlesController < ApplicationController
-  before_action :current_user!, only: :show
+  before_action :current_user!, only: [:index, :show]
   before_action :update_visits!, only: :show
+
+  def index
+    @articles = @_current_user.articles.where(status: "Published").order("updated_at DESC")
+    render
+  end
 
   def show
     render
