+++ app/javascript/src/apis/admin/index.js
@@ -1,0 +1,6 @@
+import articlesApi from "./articles";
+import categoriesApi from "./categories";
+import organizationsApi from "./organizations";
+import redirectionsApi from "./redirections";
+
+export { articlesApi, categoriesApi, redirectionsApi, organizationsApi };
