+++ app/javascript/src/apis/admin/redirections.js
