--- app/javascript/src/apis/index.js
@@ -1,14 +1,0 @@
-import articlesApi from "apis/articles";
-import { setAuthHeaders, registerIntercepts } from "apis/axios";
-import categoriesApi from "apis/categories";
-import organizationsApi from "apis/organizations";
-import redirectionsApi from "apis/redirections";
-
-export {
-  articlesApi,
-  setAuthHeaders,
-  categoriesApi,
-  redirectionsApi,
-  organizationsApi,
-  registerIntercepts,
-};
