+++ app/javascript/src/apis/public/categories.js
@@ -1,0 +1,9 @@
+import axios from "axios";
+
+const fetch = () => axios.get("/api/public/categories");
+
+const categoriesApi = {
+  fetch,
+};
+
+export default categoriesApi;
