+++ app/javascript/src/apis/public/index.js
@@ -1,0 +1,4 @@
+import articlesApi from "./articles";
+import categoriesApi from "./categories";
+
+export { articlesApi, categoriesApi };
