+++ app/javascript/src/components/Dashboard/Articles/Article/Edit.jsx
@@ -3,7 +3,7 @@ import React, { useEffect, useState } from "react";
 import { PageLoader } from "neetoui";
 import { useHistory, useParams } from "react-router-dom";
 
-import articlesApi from "apis/articles";
+import { articlesApi } from "apis/admin";
 
 import Form from "./Form";
 import VersionHistory from "./VersionHistory";
