+++ app/javascript/src/components/Dashboard/Articles/index.jsx
@@ -3,8 +3,7 @@ import React, { useState, useEffect } from "react";
 import { PageLoader } from "neetoui";
 import { Container } from "neetoui/layouts";
 
-import articlesApi from "apis/articles";
-import categoriesApi from "apis/categories";
+import { articlesApi, categoriesApi } from "apis/admin";
 
 import { ColumnsListItems } from "./constants";
 import EmptyState from "./EmptyState";
