+++ app/views/api/public/articles/index.json.jbuilder
@@ -1,0 +1,6 @@
+json.articles @articles do | article |
+  json.extract! article,
+  :id,
+  :title,
+  :slug
+end
